refactor: fix public visibility, correct imports and struct names

+++ programs/solana-challenge-application/src/constants/mod.rs
@@ -1,1 +1,1 @@
-pub const ANCHOR_DISCRIMINTOR_SIZE: u8 = 8;
+pub const ANCHOR_DISCRIMINTOR_SIZE: usize = 8;
